Minio should not always run
