+++ integration-tests/src/test/java/fr/pilato/elasticsearch/crawler/fs/test/integration/elasticsearch/FsCrawlerRestIT.java
@@ -32,22 +32,27 @@ import fr.pilato.elasticsearch.crawler.fs.rest.UploadResponse;
 import fr.pilato.elasticsearch.crawler.fs.settings.FsSettings;
 import fr.pilato.elasticsearch.crawler.fs.settings.Rest;
 import fr.pilato.elasticsearch.crawler.fs.test.integration.AbstractRestITCase;
+import io.minio.MakeBucketArgs;
+import io.minio.MinioClient;
+import io.minio.PutObjectArgs;
+import io.minio.errors.MinioException;
 import jakarta.ws.rs.core.MediaType;
+import org.apache.logging.log4j.Logger;
 import org.glassfish.jersey.media.multipart.FormDataMultiPart;
 import org.glassfish.jersey.media.multipart.file.FileDataBodyPart;
 import org.junit.Test;
+import org.testcontainers.containers.MinIOContainer;
 import org.testcontainers.containers.NginxContainer;
 import org.testcontainers.containers.wait.strategy.HttpWaitStrategy;
 import org.testcontainers.utility.MountableFile;
 
 import java.io.IOException;
 import java.net.URL;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.nio.file.*;
+import java.nio.file.attribute.BasicFileAttributes;
+import java.security.InvalidKeyException;
+import java.security.NoSuchAlgorithmException;
+import java.util.*;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Predicate;
 
@@ -369,58 +374,116 @@ public class FsCrawlerRestIT extends AbstractRestITCase {
 
     @Test
     public void testUploadDocumentWithS3Plugin() throws Exception {
-        // We try with a document that does not exist
-        String json = "{\n" +
-                "  \"type\": \"s3\",\n" +
-                "  \"s3\": {\n" +
-                "    \"url\": \"" + s3Url +"\",\n" +
-                "    \"bucket\": \"documents\",\n" +
-                "    \"object\": \"foobar/foobar.txt\",\n" +
-                "    \"access_key\": \"" + s3Username + "\",\n" +
-                "    \"secret_key\": \"" + s3Password + "\" \n" +
-                "  }\n" +
-                "}";
+        logger.info("Starting Minio");
 
-        UploadResponse uploadResponse = post(target, "/_document", json, UploadResponse.class);
-        assertThat(uploadResponse.isOk(), is(false));
-        assertThat(uploadResponse.getMessage(), containsString("FsCrawlerIllegalConfigurationException"));
-        assertThat(uploadResponse.getMessage(), containsString("The specified key does not exist"));
+        try (MinIOContainer container = new MinIOContainer("minio/minio")) {
+            container.start();
+            String s3Url = container.getS3URL();
+            String s3Username = container.getUserName();
+            String s3Password = container.getPassword();
+            logger.info("Minio started on {} with username {} and password {}. Console running at {}",
+                    s3Url, s3Username, s3Password,
+                    String.format("http://%s:%s", container.getHost(), container.getMappedPort(9001)));
 
-        // We try with an existing document
-        json = "{\n" +
-                "  \"type\": \"s3\",\n" +
-                "  \"s3\": {\n" +
-                "    \"url\": \"" + s3Url +"\",\n" +
-                "    \"bucket\": \"documents\",\n" +
-                "    \"object\": \"roottxtfile.txt\",\n" +
-                "    \"access_key\": \"" + s3Username + "\",\n" +
-                "    \"secret_key\": \"" + s3Password + "\" \n" +
-                "  }\n" +
-                "}";
-        uploadResponse = post(target, "/_document", json, UploadResponse.class);
-        assertThat(uploadResponse.isOk(), is(true));
+            // Upload all files to Minio
+            String bucket = "documents";
+            logger.debug("  --> Copying test resources from [{}] to Minio [{}] bucket", DEFAULT_RESOURCES, bucket);
 
-        // We wait until we have our document
-        ESSearchResponse response = countTestHelper(new ESSearchRequest().withIndex(getCrawlerName()), 1L, null);
-        assertThat(JsonPath.read(response.getHits().get(0).getSource(), "$.file.filename"), is("roottxtfile.txt"));
-        assertThat(JsonPath.read(response.getHits().get(0).getSource(), "$.file.filesize"), is (12230));
-        assertThat(JsonPath.read(response.getHits().get(0).getSource(), "$.content"), containsString ("Nihil est enim virtute amabilius"));
+            try (MinioClient minioClient = MinioClient.builder()
+                    .endpoint(s3Url)
+                    .credentials(s3Username, s3Password)
+                    .build()) {
+                minioClient.makeBucket(MakeBucketArgs.builder().bucket(bucket).build());
+                Files.walkFileTree(DEFAULT_RESOURCES, EnumSet.of(FileVisitOption.FOLLOW_LINKS), Integer.MAX_VALUE,
+                        new InternalFileVisitor(bucket, minioClient, logger));
+                logger.debug("  --> Test resources ready in [{}]", s3Url);
+            }
+
+            // We try with a document that does not exist
+            String json = "{\n" +
+                    "  \"type\": \"s3\",\n" +
+                    "  \"s3\": {\n" +
+                    "    \"url\": \"" + s3Url +"\",\n" +
+                    "    \"bucket\": \"documents\",\n" +
+                    "    \"object\": \"foobar/foobar.txt\",\n" +
+                    "    \"access_key\": \"" + s3Username + "\",\n" +
+                    "    \"secret_key\": \"" + s3Password + "\" \n" +
+                    "  }\n" +
+                    "}";
+
+            UploadResponse uploadResponse = post(target, "/_document", json, UploadResponse.class);
+            assertThat(uploadResponse.isOk(), is(false));
+            assertThat(uploadResponse.getMessage(), containsString("FsCrawlerIllegalConfigurationException"));
+            assertThat(uploadResponse.getMessage(), containsString("The specified key does not exist"));
+
+            // We try with an existing document
+            json = "{\n" +
+                    "  \"type\": \"s3\",\n" +
+                    "  \"s3\": {\n" +
+                    "    \"url\": \"" + s3Url +"\",\n" +
+                    "    \"bucket\": \"documents\",\n" +
+                    "    \"object\": \"roottxtfile.txt\",\n" +
+                    "    \"access_key\": \"" + s3Username + "\",\n" +
+                    "    \"secret_key\": \"" + s3Password + "\" \n" +
+                    "  }\n" +
+                    "}";
+            uploadResponse = post(target, "/_document", json, UploadResponse.class);
+            assertThat(uploadResponse.isOk(), is(true));
+
+            // We wait until we have our document
+            ESSearchResponse response = countTestHelper(new ESSearchRequest().withIndex(getCrawlerName()), 1L, null);
+            assertThat(JsonPath.read(response.getHits().get(0).getSource(), "$.file.filename"), is("roottxtfile.txt"));
+            assertThat(JsonPath.read(response.getHits().get(0).getSource(), "$.file.filesize"), is (12230));
+            assertThat(JsonPath.read(response.getHits().get(0).getSource(), "$.content"), containsString ("Nihil est enim virtute amabilius"));
+        }
     }
 
-    private final static String text = "Hello Foo world!";
+    private static class InternalFileVisitor extends SimpleFileVisitor<Path> {
+        private final String bucket;
+        private final MinioClient minioClient;
+        private final Logger logger;
+
+        public InternalFileVisitor(String bucket, MinioClient minioClient, Logger logger) {
+            this.bucket = bucket;
+            this.minioClient = minioClient;
+            this.logger = logger;
+        }
+
+        @Override
+        public FileVisitResult preVisitDirectory(Path dir, BasicFileAttributes attrs) throws IOException {
+            logger.trace("  --> Creating dir [{}] in [{}]", dir, bucket);
+            return FileVisitResult.CONTINUE;
+        }
+
+        @Override
+        public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
+            logger.trace("  --> Copying [{}] to [{}]", file, bucket);
+            try {
+                minioClient.putObject(PutObjectArgs.builder()
+                        .bucket(bucket)
+                        .object(file.getFileName().toString())
+                        .stream(Files.newInputStream(file), file.toFile().length(), -1)
+                        .build());
+            } catch (MinioException | InvalidKeyException | NoSuchAlgorithmException e) {
+                throw new IOException(e);
+            }
+            return FileVisitResult.CONTINUE;
+        }
+    }
 
     private void createFile(Path root, String objectName, String object) throws IOException {
-        staticLogger.debug("Create fake content [{}]; [{}]", objectName, object);
+        logger.debug("Create fake content [{}]; [{}]", objectName, object);
         Path file = root.resolve(objectName);
         Files.writeString(file, object);
     }
 
     @Test
     public void testUploadDocumentWithHttpPlugin() throws Exception {
-        staticLogger.info("Starting Nginx from {}", rootTmpDir);
+        logger.info("Starting Nginx from {}", rootTmpDir);
         Path nginxRoot = rootTmpDir.resolve("nginx-root");
         Files.createDirectory(nginxRoot);
         Files.writeString(nginxRoot.resolve("index.html"), "<html><body>Hello World!</body></html>");
+        String text = "Hello Foo world!";
         createFile(nginxRoot, "foo.txt", text);
         createFile(nginxRoot, "bar.txt", "This one should be ignored.");
 
@@ -429,7 +492,7 @@ public class FsCrawlerRestIT extends AbstractRestITCase {
             container.start();
             container.copyFileToContainer(MountableFile.forHostPath(nginxRoot), "/usr/share/nginx/html");
             URL url = container.getBaseUrl("http", 80);
-            staticLogger.info("Nginx started on {}.", url);
+            logger.info("Nginx started on {}.", url);
 
             // We try with a document that does not exist
             String json = "{\n" +
